fix: contract transactor key is now an externally provisioned secret (#23)
Needs https://github.com/storacha/storoku/pull/31 to land first.

Use the support for externally provisioned secrets in storoku v0.5.2 to
make the `CONTRACT_TRANSACTOR_KEY` secret external. It's value will be
removed from the repo and we will be able to manage it directly in AWS
Secrets Manager.

I confirmed this is working as expected in a test on `forge-prod`.